+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.4.0] - 2020-06-15
 
 - Project renamed from Porygon to Superhighway
 - Support for arbitrary MIME types for `Accept` and `Content-Type` headers using `ContentType.custom`
@@ -31,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release of the library
 
-[unreleased]: https://github.com/rhysforyou/Superhighway/compare/0.3.0...HEAD
+[v0.4.0]: https://github.com/rhysforyou/Superhighway/compare/0.3.0...0.4.0
 [v0.3.0]: https://github.com/rhysforyou/Superhighway/compare/0.2.0...0.3.0
 [v0.2.0]: https://github.com/rhysforyou/Superhighway/compare/0.1.0...0.2.0
 [v0.1.0]: https://github.com/rhysforyou/Superhighway/releases/tag/0.1.0